feat(native-rust): sync test redistribution and annotation placement fixes from unreviewed
